requirements.txt
amqp==5.1.1
asgiref==3.6.0
asttokens==2.2.1
backcall==0.2.0
billiard==4.1.0
boto3==1.26.103
botocore==1.29.103
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==40.0.1
cssselect==1.2.0
decorator==5.1.1
defusedxml==0.7.1
Django==4.1.7
django-cors-headers==3.14.0
django-dynamic-fixture==3.1.2
django-extensions==3.2.1
django-storages==1.13.2
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
docutils==0.19
executing==1.2.0
future==0.18.3
idna==3.4
ipdb==0.13.13
ipython==8.12.0
ipython-genutils==0.2.0
itypes==1.2.0
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
kombu==5.2.4
lxml==4.9.2
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
oauthlib==3.2.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
prompt-toolkit==3.0.38
psycopg2==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.14.0
PyJWT==2.6.0
pyquery==2.0.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3
requests==2.28.2
requests-oauthlib==1.3.1
s3transfer==0.6.0
sentry-sdk==1.18.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.4.1
sqlparse==0.4.3
stack-data==0.6.2
traitlets==5.9.0
uritemplate==4.1.1
urllib3==1.26.15
vine==5.0.0
wcwidth==0.2.6
